Refactor arrangement of file
