+++ src/main/java/application/product/ProductCategory.java
@@ -1,4 +1,4 @@
-package application;
+package application.product;
 
 public enum ProductCategory {
 	GROCERIES,
